# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API GIMA",
        "description": "Documentación de la API para el proyecto GIMA (Gestión de Inventario y Mantenimiento)",
        "contact": {
            "email": "admin@gima.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Servidor API Principal"
        }
    ],
    "paths": {
...
        "/api/catalogo/activos": {
            "get": {
                "tags": [
                    "Catálogo - Activos"
                ],
                "summary": "Listar activos",
                "operationId": "2a533e429e352803ee3257b7ba3a8057",
                "responses": {
                    "200": {
                        "description": "Lista de activos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Activo"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Catálogo - Activos"
                ],
                "summary": "Crear activo",
                "operationId": "7a746ef7d5ea553a4fac5c422b4f455c",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "articulo_id": {
                                        "type": "integer"
                                    },
                                    "ubicacion_id": {
                                        "type": "integer"
                                    },
                                    "estado": {
                                        "type": "string"
                                    },
                                    "valor": {
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Activo creado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Activo"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/catalogo/activos/{id}": {
            "get": {
                "tags": [
                    "Catálogo - Activos"
                ],
                "summary": "Ver activo",
                "operationId": "53cad12ec9f7dc36b1e7149f1ca28d58",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Activo",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Activo"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Catálogo - Activos"
                ],
                "summary": "Actualizar activo",
                "operationId": "b4e7b46e00d842db9e747ab12573c572",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "articulo_id": {
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "ubicacion_id": {
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "estado": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "valor": {
                                        "type": "number",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Activo actualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Activo"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Catálogo - Activos"
                ],
                "summary": "Eliminar activo",
                "operationId": "fcfcf463d3a40a34905888aa60b6eead",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Eliminado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/catalogo/activos/por-categoria": {
            "get": {
                "tags": [
                    "Catálogo - Activos"
                ],
                "summary": "Listar activos por categoría",
                "operationId": "a08ae855c11092445e30a255e88d49ce",
                "parameters": [
                    {
                        "name": "tipo",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Activos filtrados",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Activo"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Activo": {
                "title": "Activo",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "articulo_id": {
                        "type": "integer"
                    },
                    "ubicacion_id": {
                        "type": "integer"
                    },
                    "estado": {
                        "type": "string"
                    },
                    "valor": {
                        "type": "number"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}